pyproject.toml
[project]
name = "mini-rag-qa"
version = "0.1.0"
description = "RAG-based QA over PostgreSQL 16 documentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.9.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "unstructured>=0.16.0",
    "scikit-learn>=1.5.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
]
